working no conflicts libft, still need to clean useless functions in printf and gnl than can be get on libft

+++ sources/libft/INT_STR/ft_itoa.c
@@ -30,7 +30,7 @@ static int		get_lenght(long int n)
 	return (len);
 }
 
-char			*ft_itoa(int n)
+char			*ft_itoa(long int n)
 {
 	long int	num;
 	char		*number;
